Add check for current user process only.

+++ Cleanup.vbs
@@ -1,9 +1,15 @@
+Set objNetwork = Wscript.CreateObject("Wscript.Network")
+currUser = LCase(objNetwork.UserName)
 Const strComputer = "." 
   Dim objWMIService, colProcessList
   Set objWMIService = GetObject("winmgmts:" & "{impersonationLevel=impersonate}!\\" & strComputer & "\root\cimv2")
   Set colProcessList = objWMIService.ExecQuery("SELECT * FROM Win32_Process WHERE Name = 'ssh-agent.exe'")
   For Each objProcess in colProcessList 
-    objProcess.Terminate() 
+	If objProcess.GetOwner ( User, Domain ) = 0 Then
+		If LCase(User) = currUser Then
+			objProcess.Terminate() 
+		end if
+	end if
   Next
